# Speedup proptests for Chain struct in zebra-state