## Motivation

Improve `proptest` performance for `Chain` struct because currently only 1 case set in `ProptestConfig`.

## Solution

Instead of generating a new chain for each `proptest` case generate it only once and searching only on the indexes.

Without PR, 1 case:
```
test service::non_finalized_state::chain::tests::forked_equals_pushed ... ok
test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 25 filtered out; finished in 3.94s
test service::non_finalized_state::chain::tests::finalized_equals_pushed ... ok
test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 25 filtered out; finished in 4.26s
```

With PR, 32 cases:
```
test service::non_finalized_state::chain::tests::forked_equals_pushed ... ok
test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 25 filtered out; finished in 8.64s
test service::non_finalized_state::chain::tests::finalized_equals_pushed ... ok
test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 25 filtered out; finished in 7.45s
```

Tested on `i5-8250U`.

## Review

## Related Issues

Closes https://github.com/ZcashFoundation/zebra/issues/1112.

## Follow Up Work

Nope.