@fanatid 2021-04-15
hah, tested with `cargo clippy -p zebra-chain` locally instead `-p zebra-state`